requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
numpy==1.25.0
pandas==2.0.3
python-dotenv==1.0.0
celery[mongodb,rabbitmq]==5.3.5
Flask-Cors==4.0.0
flatbuffers==23.5.26
pyyaml==6.0.1
requests==2.31.0